// gen/locate2u/models/ClubBulkUploadItem.ts
export type ClubBulkUploadItem = {
  /**
   * @type string | undefined
   */
  name?: string | null;
  /**
   * @type string | undefined
   */

  image?: string | null;

  /**
   * @type string | undefined
   */

  sport?: string | null;
  /**
   * @type string | undefined
   */

  club?: string | null;
  /**
   * @type string | undefined
   */

  details?: string | null;
  /**
   * @type number | undefined
   */

  positionId?: number | null;
};